move exception package from headlong.util to headlong
