+++ src/main/java/com/esaulpaugh/headlong/exception/RecoverableDecodeException.java
@@ -13,23 +13,24 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.esaulpaugh.headlong.util.exception;
+package com.esaulpaugh.headlong.exception;
 
 /**
- * Indicates a failure to decode an item that is unrecoverably malformed.
+ * Indicates a failure to decode an RLP item due to a short buffer, potentially because the item has not finished
+ * streaming, i.e. it is a prefix of some hypothetical, valid longer item.
  */
-public final class UnrecoverableDecodeException extends DecodeException {
+public final class RecoverableDecodeException extends DecodeException {
 
-    public UnrecoverableDecodeException(String msg) {
+    public RecoverableDecodeException(String msg) {
         super(msg);
     }
 
-    public UnrecoverableDecodeException(Throwable cause) {
+    public RecoverableDecodeException(Throwable cause) {
         super(cause);
     }
 
     @Override
     public boolean isRecoverable() {
-        return false;
+        return true;
     }
 }
