+++ src/main/java/com/esaulpaugh/headlong/exception/UnrecoverableDecodeException.java
@@ -13,24 +13,23 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.esaulpaugh.headlong.util.exception;
+package com.esaulpaugh.headlong.exception;
 
 /**
- * Indicates a failure to decode an RLP item due to a short buffer, potentially because the item has not finished
- * streaming, i.e. it is a prefix of some hypothetical, valid longer item.
+ * Indicates a failure to decode an item that is unrecoverably malformed.
  */
-public final class RecoverableDecodeException extends DecodeException {
+public final class UnrecoverableDecodeException extends DecodeException {
 
-    public RecoverableDecodeException(String msg) {
+    public UnrecoverableDecodeException(String msg) {
         super(msg);
     }
 
-    public RecoverableDecodeException(Throwable cause) {
+    public UnrecoverableDecodeException(Throwable cause) {
         super(cause);
     }
 
     @Override
     public boolean isRecoverable() {
-        return true;
+        return false;
     }
 }
