+++ src/main/java/com/esaulpaugh/headlong/rlp/RLPItem.java
@@ -15,9 +15,9 @@
 */
 package com.esaulpaugh.headlong.rlp;
 
-import com.esaulpaugh.headlong.util.exception.DecodeException;
-import com.esaulpaugh.headlong.util.exception.RecoverableDecodeException;
-import com.esaulpaugh.headlong.util.exception.UnrecoverableDecodeException;
+import com.esaulpaugh.headlong.exception.DecodeException;
+import com.esaulpaugh.headlong.exception.RecoverableDecodeException;
+import com.esaulpaugh.headlong.exception.UnrecoverableDecodeException;
 import com.esaulpaugh.headlong.rlp.util.FloatingPoint;
 import com.esaulpaugh.headlong.util.Integers;
 import com.esaulpaugh.headlong.rlp.util.Notation;
