+++ src/main/java/com/esaulpaugh/headlong/rlp/util/FloatingPoint.java
@@ -15,7 +15,7 @@
 */
 package com.esaulpaugh.headlong.rlp.util;
 
-import com.esaulpaugh.headlong.util.exception.DecodeException;
+import com.esaulpaugh.headlong.exception.DecodeException;
 import com.esaulpaugh.headlong.util.Integers;
 
 import java.math.BigDecimal;
