+++ src/test/java/com/esaulpaugh/headlong/TestUtils.java
@@ -18,7 +18,7 @@ package com.esaulpaugh.headlong;
 import com.esaulpaugh.headlong.util.Integers;
 import com.esaulpaugh.headlong.util.FastHex;
 import com.esaulpaugh.headlong.util.Strings;
-import com.esaulpaugh.headlong.util.exception.DecodeException;
+import com.esaulpaugh.headlong.exception.DecodeException;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
