+++ src/test/java/com/esaulpaugh/headlong/rlp/example/RLPStudent.java
@@ -21,7 +21,7 @@ import com.esaulpaugh.headlong.rlp.RLPIterator;
 import com.esaulpaugh.headlong.rlp.util.FloatingPoint;
 import com.esaulpaugh.headlong.util.Integers;
 import com.esaulpaugh.headlong.util.Strings;
-import com.esaulpaugh.headlong.util.exception.DecodeException;
+import com.esaulpaugh.headlong.exception.DecodeException;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
