+++ src/test/java/com/esaulpaugh/headlong/rlp/util/IntegersTest.java
@@ -18,7 +18,7 @@ package com.esaulpaugh.headlong.rlp.util;
 import com.esaulpaugh.headlong.TestUtils;
 import com.esaulpaugh.headlong.util.Integers;
 import com.esaulpaugh.headlong.util.Strings;
-import com.esaulpaugh.headlong.util.exception.DecodeException;
+import com.esaulpaugh.headlong.exception.DecodeException;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigInteger;
